product analytics data refactorization
