+++ src/Form/Modifier/ProductFormModifier.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Oksydan\IsFavoriteProducts\Form\Modifier;
 
 use Oksydan\IsFavoriteProducts\Form\Type\ProductFavoriteType;
-use Oksydan\IsFavoriteProducts\View\Admin\RenderAdminProductFavoriteStats;
 use PrestaShopBundle\Form\FormBuilderModifier;
 use Symfony\Component\Form\FormBuilderInterface;
 
@@ -17,19 +16,12 @@ class ProductFormModifier
     private FormBuilderModifier $formBuilderModifier;
 
     /**
-     * @var RenderAdminProductFavoriteStats
-     */
-    private RenderAdminProductFavoriteStats $renderAdminProductFavoriteStats;
-
-    /**
      * @param FormBuilderModifier $formBuilderModifier
      */
     public function __construct(
-        FormBuilderModifier $formBuilderModifier,
-        RenderAdminProductFavoriteStats $renderAdminProductFavoriteStats
+        FormBuilderModifier $formBuilderModifier
     ) {
         $this->formBuilderModifier = $formBuilderModifier;
-        $this->renderAdminProductFavoriteStats = $renderAdminProductFavoriteStats;
     }
 
     /**
@@ -46,12 +38,8 @@ class ProductFormModifier
             'favorite_products',
             ProductFavoriteType::class,
             [
-                'data' => [
-                    'html_content' => [
-                        'content' => $this->renderAdminProductFavoriteStats->render($productId),
-                    ],
-                ],
-            ]
+                'product_id' => $productId,
+            ],
         );
     }
 }
