--- src/View/Admin/RenderInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Oksydan\IsFavoriteProducts\View\Admin;
-
-interface RenderInterface
-{
-    public function render(int $id): string;
-}
