+++ views/templates/admin/FormTheme/product_form_theme.html.twig
@@ -1,0 +1,7 @@
+{% extends '@PrestaShop/Admin/TwigTemplateForm/prestashop_ui_kit_base.html.twig' %}
+
+{% block product_favorite_widget %}
+  {{ include('@Modules/is_favoriteproducts/views/templates/admin/product/stats.html.twig', {
+    'productStats': productStats,
+  }) }}
+{% endblock %}
