+++ views/templates/admin/hook/displayAdminProductsExtra.html.twig
@@ -1,0 +1,7 @@
+
+{% block content %}
+  {{ form_start(form) }}
+    {{ form_widget(form) }}
+    {{ form_rest(form) }}
+  {{ form_end(form) }}
+{% endblock %}
